Clean up and refactor
LayerSchemaCompiler does schema-compile-time checking and constructs a LayerFactory

LayerRuntimeCompiler does
 runtime compiling of the construct, parameter/var setting,
 result/var getting structures.

 Both LayerSchemaCompiler and
 LayerFactory invoke the LayerRuntimeCompiler, in the former case to do checking.

 LayerRegistry is gone, merged into LayerFactory
