+++ daffodil-runtime1/src/main/scala/org/apache/daffodil/runtime1/layers/api/ChecksumLayer.java
@@ -24,9 +24,8 @@ import java.nio.ByteBuffer;
  * A checksum layer computes some sort of integer value from a region of the data stream. The term checksum is
  * used generically here to subsume all sorts of CRCs, check digits, data hash, and digest calculations.
  * <p/>
- * This base is suitable only for checksums computed over small sections of data, not large data streams or whole files.
- * <p/>
- * The entire region of data the checksum is being computed over, will be pulled into a byte buffer in memory.
+ * This base is suitable only for checksums computed over small sections of data, not large data streams or whole large
+ * files. The entire region of data the checksum is being computed over, will be pulled into a byte buffer in memory.
  * <p/>
  * The resulting checksum is the return value of the compute method.
  * <p/>
@@ -41,11 +40,32 @@ import java.nio.ByteBuffer;
  * <p/>
  * This will be called automatically to retrieve the integer value that was returned from the `compute` method, and
  * the DFDL variable 'digest' will be assigned that value.
+ * <p/>
+ * The derived class implementing a checksum layer must call
+ * <p/>
+ *     setLength(len) // sets the length in bytes
+ * <p/>
+ * to specify the length of the data region in bytes. Normally this would be called from the layer's implementation of
+ * the
+ * <p/>
+ *     void setLayerVariableParameters(...) { }
+ * <p/>
+ * method, which, if defined, is called with arguments populated from DFDL variables with the same name (and compatible type)
+ * defined in a DFDL schema with the Layer's target namespace. So, for example if a checksum layer needs to
+ * receive a parameter from a DFDL variable named "layerEncoding", the setter would be:
+ * <p/>
+ *     void setLayerVariableParameters(String layerEncoding) {
+ *         this.layerEncoding = layerEncoding;
+ *     }
+ * <p/>
+ * Beside initializing local members, this setter is also an initializer for the layer class instance. Any exception
+ * thrown becomes a Schema Definition Error. If there are no parameter variables, then this setter, with no arguments,
+ * can be used purely for initialization.
  */
 public abstract class ChecksumLayer extends ChecksumLayerBase {
 
-  public ChecksumLayer(String layerName, String layerNamespace) {
-    super(layerName, layerNamespace);
+  public ChecksumLayer(String layerName, String layerTargetNamespace) {
+    super(layerName, layerTargetNamespace);
   }
 
   /**
